reduce the number of shots for testing and add class definitons on conftest
