+++ tests/conftest.py
@@ -127,6 +127,9 @@ LightningException = None
 
 if device_name == "lightning.kokkos":
     from pennylane_lightning.lightning_kokkos import LightningKokkos as LightningDevice
+    from pennylane_lightning.lightning_kokkos._state_vector import LightningKokkosStateVector as LightningStateVector
+    from pennylane_lightning.lightning_kokkos._measurements import LightningKokkosMeasurements as LightningMeasurements
+    from pennylane_lightning.lightning_kokkos._adjoint_jacobian import LightningKokkosAdjointJacobian as LightningAdjointJacobian
 
     if hasattr(pennylane_lightning, "lightning_kokkos_ops"):
         import pennylane_lightning.lightning_kokkos_ops as lightning_ops
@@ -145,6 +148,12 @@ elif device_name == "lightning.tensor":
         from pennylane_lightning.lightning_tensor_ops import LightningException
 else:
     from pennylane_lightning.lightning_qubit import LightningQubit as LightningDevice
+    from pennylane_lightning.lightning_qubit._state_vector import LightningStateVector
+    from pennylane_lightning.lightning_qubit._measurements import LightningMeasurements
+    from pennylane_lightning.lightning_qubit._adjoint_jacobian import LightningAdjointJacobian
+    
+    
+    
 
     if hasattr(pennylane_lightning, "lightning_qubit_ops"):
         import pennylane_lightning.lightning_qubit_ops as lightning_ops
@@ -163,6 +172,18 @@ def qubit_device(request):
         return qml.device(device_name, wires=wires, shots=shots, c_dtype=request.param)
 
     return _device
+
+
+# General LightningStateVector fixture, for any number of wires.
+@pytest.fixture(
+    scope="function",
+    params=[np.complex64, np.complex128],
+)
+def lightning_sv(request):
+    def _statevector(num_wires):
+        return LightningStateVector(num_wires=num_wires, dtype=request.param)
+
+    return _statevector
 
 
 #######################################################################
