+++ tests/lightning_qubit/test_jacobian_method.py
@@ -16,12 +16,11 @@
 import numpy as np
 import pennylane as qml
 import pytest
-from conftest import PHI, THETA, LightningDevice, device_name  # tested device
+from conftest import PHI, THETA, LightningDevice, LightningStateVector, device_name  # tested device
 from pennylane.devices import DefaultExecutionConfig, DefaultQubit, ExecutionConfig
 from pennylane.tape import QuantumScript
 
 if device_name == "lightning.qubit":
-    from pennylane_lightning.lightning_qubit._state_vector import LightningStateVector
     from pennylane_lightning.lightning_qubit.lightning_qubit import (
         jacobian,
         simulate,
@@ -30,11 +29,7 @@ if device_name == "lightning.qubit":
         vjp,
     )
 
-
 if device_name == "lightning.kokkos":
-    from pennylane_lightning.lightning_kokkos._state_vector import (
-        LightningKokkosStateVector as LightningStateVector,
-    )
     from pennylane_lightning.lightning_kokkos.lightning_kokkos import (
         jacobian,
         simulate,
@@ -42,7 +37,6 @@ if device_name == "lightning.kokkos":
         simulate_and_vjp,
         vjp,
     )
-
 
 if not LightningDevice._new_API:
     pytest.skip(
@@ -52,18 +46,6 @@ if not LightningDevice._new_API:
 
 if not LightningDevice._CPP_BINARY_AVAILABLE:
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)
-
-
-# General LightningStateVector fixture, for any number of wires.
-@pytest.fixture(
-    scope="module",
-    params=[np.complex64, np.complex128],
-)
-def lightning_sv(request):
-    def _statevector(num_wires):
-        return LightningStateVector(num_wires=num_wires, dtype=request.param)
-
-    return _statevector
 
 
 class TestJacobian:
